backend/requirements.txt
django>=4.2.0,<5.3.0
djangorestframework>=3.14.0,<3.17.0
django-cors-headers>=4.0.0,<4.8.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<1.2.0
djangorestframework-simplejwt>=5.2.0,<5.4.0